Inclusion of Events
The base for the event system has been implemented. There is a subclass in the game manager with the function for each event and how to call them.
So far I got the events related to robbing written down. I still need to write the ones for the other categories.
(I NEED TO TEST ALL OF THESE ONCE THEY ARE DONE)

Also I just noticed the activity class and the district class only included 5 districts but there are 6, so I fixed that.